+++ src/Sentry/SentryEvent.cs
@@ -242,6 +242,14 @@ public sealed class SentryEvent : IEventLike, IJsonSerializable, IHasDistributio
     public void UnsetTag(string key) =>
         (_tags ??= new Dictionary<string, string>()).Remove(key);
 
+    internal void Redact()
+    {
+        foreach (var breadcrumb in Breadcrumbs)
+        {
+            breadcrumb.Redact();
+        }
+    }
+
     /// <inheritdoc />
     public void WriteTo(Utf8JsonWriter writer, IDiagnosticLogger? logger)
     {